+++ src/Sulu/Bundle/RouteBundle/Content/Type/RouteContentType.php
@@ -1,0 +1,43 @@
+<?php
+
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\RouteBundle\Content\Type;
+
+use Sulu\Component\Content\SimpleContentType;
+
+/**
+ * Provides simple route edit.
+ */
+class RouteContentType extends SimpleContentType
+{
+    /**
+     * @var string
+     */
+    private $template;
+
+    /**
+     * @param string $template
+     */
+    public function __construct($template)
+    {
+        parent::__construct('Route', '');
+
+        $this->template = $template;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getTemplate()
+    {
+        return $this->template;
+    }
+}